Bosses

See merge request COMP2511/22T2/teams/H18A_BROWNIE/project!63

+++ src/main/java/dungeonmania/entities/collectableEntities/Bomb.java
@@ -19,7 +19,7 @@ public class Bomb extends Item {
     }
 
     public int getBombRadius() {
-        return JSONConfig.getConfig("bomb_radius");
+        return (int) JSONConfig.getConfig("bomb_radius");
     }
 
     public void explode(DungeonMap map) {
